turnkey_web/requirements.txt
# Only for turnkey_web
Django>=3.2,<3.3
django-compressor>=2.4,<2.5
djangorestframework>=3.12,<3.13
djangorestframework-filters==1.0.0.dev0
django-extensions>=3.1,<3.2
mysqlclient>=2.0,<2.1
